requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
sqlalchemy>=2.0.0
groq>=0.11.0
pillow>=10.3.0
python-dotenv>=1.0.0
pytest>=8.0.0
httpx>=0.27.0
